+++ src/pdm_conda/models/environment.py
@@ -5,15 +5,15 @@ from copy import copy
 from pathlib import Path
 from typing import cast
 
-from pdm.exceptions import NoPythonVersion, ProjectError
+from pdm.exceptions import ProjectError
 from pdm.models.environment import Environment, PrefixEnvironment
 from pdm.models.requirements import Requirement
+from pdm.models.specifiers import PySpecSet
 from pdm.models.working_set import WorkingSet
 from pdm.project import Project
 
 from pdm_conda.conda import conda_list, conda_search
 from pdm_conda.mapping import pypi_to_conda
-from pdm_conda.models.candidates import CondaCandidate
 from pdm_conda.project import CondaProject
 from pdm_conda.utils import normalize_name
 
@@ -31,7 +31,7 @@ class CondaEnvironment(Environment):
         super().__init__(project)
         self.project = cast(CondaProject, project)
         self._python_dependencies: dict[str, Requirement] | None = None
-        self._python_candidate: CondaCandidate | None = None
+        self.python_requires &= PySpecSet(f"=={self.interpreter.version}")
 
     @property
     def packages_path(self) -> Path:
@@ -54,15 +54,6 @@ class CondaEnvironment(Environment):
         return working_set
 
     @property
-    def python_candidate(self) -> CondaCandidate:
-        if self._python_candidate is None:
-            python_package = conda_list(self.project).get("python", None)
-            if python_package is None:
-                raise NoPythonVersion("No python found in Conda environment.")
-            self._python_candidate = conda_search(python_package.as_line().replace(" ", "="), self.project)[0]
-        return self._python_candidate
-
-    @property
     def python_dependencies(self) -> dict[str, Requirement]:
         if self._python_dependencies is None:
             self._python_dependencies = dict()
@@ -71,7 +62,7 @@ class CondaEnvironment(Environment):
                 if name not in packages and name not in dependencies:
                     return
                 package = packages[name].as_line().replace(" ", "=")
-                candidate = conda_search(package, self.project)[0]
+                candidate = conda_search(self.project, package)[0]
                 dependencies[name] = candidate.req
                 for d in candidate.dependencies:
                     load_dependencies(d.name, packages, dependencies)